moved files to Megathread.js/.css
