+++ front-end/src/Megathread.css
@@ -1,34 +1,34 @@
-.App {
-    text-align: center;
+.Megathread {
+    text-align: left;
   }
   
-  .App-logo {
+  .Megathread-logo {
     height: 40vmin;
     pointer-events: none;
   }
   
   @media (prefers-reduced-motion: no-preference) {
-    .App-logo {
-      animation: App-logo-spin infinite 20s linear;
+    .Megathread-logo {
+      animation: Megathread-logo-spin infinite 20s linear;
     }
   }
   
-  .App-header {
-    background-color: #282c34;
+  .Megathread-header {
+    background-color: #5C667A;
     min-height: 100vh;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
     font-size: calc(10px + 2vmin);
-    color: white;
+    color: #282C35;
   }
   
-  .App-link {
+  .Megathread-link {
     color: #61dafb;
   }
   
-  @keyframes App-logo-spin {
+  @keyframes Megathread-logo-spin {
     from {
       transform: rotate(0deg);
     }
